added genserver API to kumbuka
